Use generic list of floats in GraphView and move the class to other package

+++ app/src/main/java/org/wikipedia/feed/mostread/MostReadArticles.java
@@ -25,14 +25,14 @@ public final class MostReadArticles extends PageSummary {
 
     public class ViewHistory {
         private Date date;
-        private int views;
+        private float views;
 
         @NonNull
         public Date getDate() {
             return date;
         }
 
-        public int getViews() {
+        public float getViews() {
             return views;
         }
     }
